@@ -1,4 +1,4 @@
-package com.mydoctorapp.my_doctor_app
+package com.doctorclient.user_doctor_client
 
 import io.flutter.embedding.android.FlutterActivity
 
