@@ -1,5 +1,5 @@
-import 'package:my_doctor_app/constant/constant.dart';
-import 'package:my_doctor_app/pages/chat/chat.dart';
+import 'package:user_doctor_client/constant/constant.dart';
+import 'package:user_doctor_client/pages/chat/chat.dart';
 import 'package:flutter/material.dart';
 import 'package:page_transition/page_transition.dart';
 
