@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:my_doctor_app/constant/constant.dart';
-import 'package:my_doctor_app/pages/lab/lab.dart';
+import 'package:user_doctor_client/constant/constant.dart';
+import 'package:user_doctor_client/pages/lab/lab.dart';
 import 'package:page_transition/page_transition.dart';
 
 class LabList extends StatefulWidget {
