@@ -4,10 +4,10 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:my_doctor_app/constant/constant.dart';
+import 'package:user_doctor_client/constant/constant.dart';
 import 'package:flutter/material.dart';
-import 'package:my_doctor_app/pages/screens.dart';
-import 'package:my_doctor_app/provider/auth_provider.dart';
+import 'package:user_doctor_client/pages/screens.dart';
+import 'package:user_doctor_client/provider/auth_provider.dart';
 import 'package:page_transition/page_transition.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:provider/provider.dart';
