@@ -1,301 +1,143 @@
-
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:my_doctor_app/constant/constant.dart';
-import 'package:my_doctor_app/pages/screens.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:razorpay_flutter/razorpay_flutter.dart';
+import 'package:user_doctor_client/constant/constant.dart';
+import 'package:user_doctor_client/pages/appointment/appointment.dart';
 
 class Payment extends StatefulWidget {
+  const Payment({Key key}) : super(key: key);
+
   @override
   _PaymentState createState() => _PaymentState();
 }
 
 class _PaymentState extends State<Payment> {
-  bool amazon = true,
-      card = false,
-      paypal = false,
-      skrill = false,
-      cashOn = false;
+  FirebaseFirestore db = FirebaseFirestore.instance;
+  Razorpay _razorpay = Razorpay();
+  var orderId;
 
-  successOrderDialog() {
+  int commision;
+
+  String paymentId;
+
+  String signature;
+  void _showDialog(
+      BuildContext context, String title, String message, Function function) {
     showDialog(
       context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        // return object of type Dialog
-        return Dialog(
-          elevation: 0.0,
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)),
-          child: Container(
-            height: 170.0,
-            padding: EdgeInsets.all(20.0),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                Container(
-                  height: 70.0,
-                  width: 70.0,
-                  alignment: Alignment.center,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(35.0),
-                    border: Border.all(color: primaryColor, width: 1.0),
-                  ),
-                  child: Icon(
-                    Icons.check,
-                    size: 40.0,
-                    color: primaryColor,
-                  ),
-                ),
-                SizedBox(
-                  height: 20.0,
-                ),
-                Text(
-                  "Success!",
-                  style: greySmallBoldTextStyle,
-                  textAlign: TextAlign.center,
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-
-    Future.delayed(const Duration(milliseconds: 3000), () {
-      setState(() {
-        Navigator.push(
-          context,
-          MaterialPageRoute(builder: (context) => BottomBar()),
-        );
-      });
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        titleSpacing: 0.0,
-        elevation: 1.0,
-        title: Text(
-          'Select Payment Method',
-          style: appBarWhiteTitleTextStyle,
-        ),
-      ),
-      bottomNavigationBar: Material(
-        elevation: 5.0,
-        child: Container(
-          color: Colors.white,
-          width: double.infinity,
-          height: 70.0,
-          padding: EdgeInsets.symmetric(horizontal: fixPadding * 2.0),
-          alignment: Alignment.center,
-          child: InkWell(
-            borderRadius: BorderRadius.circular(15.0),
-            onTap: () {
-              successOrderDialog();
-            },
-            child: Container(
-              width: double.infinity,
-              height: 50.0,
-              alignment: Alignment.center,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(15.0),
-                color: primaryColor,
-              ),
-              child: Text(
-                'Pay',
-                style: whiteColorButtonTextStyle,
-              ),
-            ),
-          ),
-        ),
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Container(
-            width: double.infinity,
-            padding: EdgeInsets.all(fixPadding * 2.0),
-            color: lightPrimaryColor,
-            child: Text(
-              'Pay: \$39',
-              style: blackHeadingTextStyle,
-            ),
-          ),
-          Expanded(
-            child: ListView(
-              children: [
-                getPaymentTile(
-                    'Pay on Visit', 'assets/payment_icon/cash_on_delivery.png'),
-                getPaymentTile(
-                    'Amazon Pay', 'assets/payment_icon/amazon_pay.png'),
-                getPaymentTile('Card', 'assets/payment_icon/card.png'),
-                getPaymentTile('PayPal', 'assets/payment_icon/paypal.png'),
-                getPaymentTile('Skrill', 'assets/payment_icon/skrill.png'),
-                Container(height: fixPadding * 2.0),
-              ],
-            ),
+      builder: (ctx) => AlertDialog(
+        title: Text(title),
+        content: Text(message),
+        actions: <Widget>[
+          TextButton(
+            onPressed: function,
+            child: Text('OK!'),
           ),
         ],
       ),
     );
   }
 
-  getPaymentTile(String title, String imgPath) {
-    return InkWell(
-      onTap: () {
-        if (title == 'Pay on Visit') {
-          setState(() {
-            cashOn = true;
-            amazon = false;
-            card = false;
-            paypal = false;
-            skrill = false;
-          });
-        } else if (title == 'Amazon Pay') {
-          setState(() {
-            cashOn = false;
-            amazon = true;
-            card = false;
-            paypal = false;
-            skrill = false;
-          });
-        } else if (title == 'Card') {
-          setState(() {
-            cashOn = false;
-            amazon = false;
-            card = true;
-            paypal = false;
-            skrill = false;
-          });
-        } else if (title == 'PayPal') {
-          setState(() {
-            cashOn = false;
-            amazon = false;
-            card = false;
-            paypal = true;
-            skrill = false;
-          });
-        } else if (title == 'Skrill') {
-          setState(() {
-            cashOn = false;
-            amazon = false;
-            card = false;
-            paypal = false;
-            skrill = true;
-          });
-        }
-      },
-      child: Container(
-        margin: EdgeInsets.only(
-            right: fixPadding * 2.0,
-            left: fixPadding * 2.0,
-            top: fixPadding * 2.0),
-        padding: EdgeInsets.all(fixPadding * 2.0),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(7.0),
-          border: Border.all(
-            width: 1.0,
-            color: (title == 'Pay on Visit')
-                ? (cashOn)
-                    ? primaryColor
-                    : Colors.grey[300]
-                : (title == 'Amazon Pay')
-                    ? (amazon)
-                        ? primaryColor
-                        : Colors.grey[300]
-                    : (title == 'Card')
-                        ? (card)
-                            ? primaryColor
-                            : Colors.grey[300]
-                        : (title == 'PayPal')
-                            ? (paypal)
-                                ? primaryColor
-                                : Colors.grey[300]
-                            : (skrill)
-                                ? primaryColor
-                                : Colors.grey[300],
-          ),
-          color: whiteColor,
+  @override
+  void initState() {
+    super.initState();
+    getCommision();
+    _razorpay.on(Razorpay.EVENT_PAYMENT_SUCCESS, _handlePaymentSuccess);
+    _razorpay.on(Razorpay.EVENT_PAYMENT_ERROR, _handlePaymentError);
+    _razorpay.on(Razorpay.EVENT_EXTERNAL_WALLET, _handleExternalWallet);
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _razorpay.clear();
+  }
+
+  void _handlePaymentSuccess(PaymentSuccessResponse response) {
+    orderId = response.orderId;
+    paymentId = response.paymentId;
+    signature = response.signature;
+    // Do something when payment succeeds
+    print("---success----${response.orderId}");
+    _showDialog(context, "Success", "Your payment Succesful", () {
+      Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(
+          builder: (context) => Appointment(),
         ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Row(
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Container(
-                  width: 50.0,
-                  child:
-                      Image.asset(imgPath, width: 50.0, fit: BoxFit.fitWidth),
-                ),
-                widthSpace,
-                Text(title, style: primaryColorHeadingTextStyle),
-              ],
-            ),
-            Container(
-              width: 20.0,
-              height: 20.0,
-              alignment: Alignment.center,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10.0),
-                border: Border.all(
-                  width: 1.5,
-                  color: (title == 'Pay on Visit')
-                      ? (cashOn)
-                          ? primaryColor
-                          : Colors.grey[300]
-                      : (title == 'Amazon Pay')
-                          ? (amazon)
-                              ? primaryColor
-                              : Colors.grey[300]
-                          : (title == 'Card')
-                              ? (card)
-                                  ? primaryColor
-                                  : Colors.grey[300]
-                              : (title == 'PayPal')
-                                  ? (paypal)
-                                      ? primaryColor
-                                      : Colors.grey[300]
-                                  : (skrill)
-                                      ? primaryColor
-                                      : Colors.grey[300],
-                ),
-              ),
-              child: Container(
-                width: 10.0,
-                height: 10.0,
-                alignment: Alignment.center,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(5.0),
-                  color: (title == 'Pay on Visit')
-                      ? (cashOn)
-                          ? primaryColor
-                          : Colors.transparent
-                      : (title == 'Amazon Pay')
-                          ? (amazon)
-                              ? primaryColor
-                              : Colors.transparent
-                          : (title == 'Card')
-                              ? (card)
-                                  ? primaryColor
-                                  : Colors.transparent
-                              : (title == 'PayPal')
-                                  ? (paypal)
-                                      ? primaryColor
-                                      : Colors.transparent
-                                  : (skrill)
-                                      ? primaryColor
-                                      : Colors.transparent,
-                ),
-              ),
-            ),
-          ],
-        ),
+      );
+    });
+
+    db.collection("extras").add({
+      "orderId": orderId,
+      "paymentId": paymentId,
+      "signature": signature,
+    }).whenComplete(() {
+      Fluttertoast.showToast(
+          msg: "Your request is received!", toastLength: Toast.LENGTH_SHORT);
+    });
+  }
+
+  void _handlePaymentError(PaymentFailureResponse response) {
+    // Do something when payment fails
+    print("---error----${response.message}");
+    _showDialog(context, "Error Occured!", response.message, () {
+      Navigator.of(context).pop();
+    });
+
+    Fluttertoast.showToast(
+        msg: "ERROR: " + response.code.toString() + " - " + response.message,
+        toastLength: Toast.LENGTH_SHORT);
+  }
+
+  void _handleExternalWallet(ExternalWalletResponse response) {
+    // Do something when an external wallet was selected
+    // _showDialog(context, "Success", "Your payment Succesful");
+
+    print("---wallet----${response.walletName}");
+    Fluttertoast.showToast(
+        msg: "EXTERNAL_WALLET: " + response.walletName,
+        toastLength: Toast.LENGTH_SHORT);
+  }
+
+  getCommision() {
+    db.collection("extras").doc("commision").get().then((value) {
+      final dat = value.data();
+      commision = dat['com'];
+      print("-----------${dat['com']}");
+    });
+  }
+
+  doPayement() async {
+    var options = {
+      'key': paymentApiKey,
+      'amount': commision + 100,
+      'name': 'Doctor Client',
+      'description': 'Payment for appointment',
+      'prefill': {'contact': '8888888888', 'email': 'test@razorpay.com'},
+      'external': {
+        'wallets': ['paytm']
+      }
+    };
+    try {
+      _razorpay.open(options);
+    } catch (e) {
+      debugPrint('Error: e');
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: ListView(
+        children: [
+          TextButton(
+              onPressed: () {
+                doPayement();
+              },
+              child: Text("Click"))
+        ],
       ),
     );
   }
